feat: noticket - Update Document __eq__/__ne__ to return NotImplemented when comparing to different classes.
Why? Short reason: We have the concept of MongoRef (a light wrapper that
holds a pointer to a MongoEngine document). When performing basic
equality checks we would like a MongoRef and Document of the same
type/id to be considered equal.

This update causes MongoEngine to return NotImplemented instead of False
when comparing against things that are not the same class. Which will
allow us to implement/allow the comparison in our MongoRef class.

https://docs.python.org/2/reference/datamodel.html

NotImplemented

This type has a single value. There is a single object with this value.
This object is accessed through the built-in name NotImplemented.
Numeric methods and rich comparison methods may return this value if
they do not implement the operation for the operands provided. (The
interpreter will then try the reflected operation, or some other
fallback, depending on the operator.) Its truth value is true.