fix: MONSTER-65 - fix for default 'id' field creation
Mainly from, this PR with one addition of setting primary_key and an
extra test.
https://github.com/MongoEngine/mongoengine/pull/961/

Internally MongoEngine maintains a mapping of field names to values in a
dictionary called `_data`. Due to an issue where the automatic id field
is created with a name attribute set to `None` it causes the id
attribute to be stored under `None` in the `_data` dictionary.

The shard key checks use this internal mapping to check if you're
updating the shardkey. These fail if your shard key contains the `id`
field.